// packages/sdk/gen/tfplugin6.ts
export enum StringKind {
  PLAIN = 0,
  MARKDOWN = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface Schema {
  /**
   * The version of the schema.
   * Schemas are versioned, so that providers can upgrade a saved resource
   * state when the schema is changed.
   */
  version: Long;
  /** Block is the top level configuration block for this schema. */
  block: Schema_Block | undefined;
}

export interface Schema_Block {
  version: Long;
  attributes: Schema_Attribute[];
  blockTypes: Schema_NestedBlock[];
  description: string;
  descriptionKind: StringKind;
  deprecated: boolean;
  deprecationMessage: string;
  computed: boolean;
}

export interface Schema_Attribute {
  name: string;
  type: Uint8Array;
  nestedType: Schema_Object | undefined;
  description: string;
  required: boolean;
  optional: boolean;
  computed: boolean;
  sensitive: boolean;
  descriptionKind: StringKind;
  deprecated: boolean;
  writeOnly: boolean;
  deprecationMessage: string;
}

export interface Schema_NestedBlock {
  typeName: string;
  block: Schema_Block | undefined;
  nesting: Schema_NestedBlock_NestingMode;
  minItems: Long;
  maxItems: Long;
}

export enum Schema_NestedBlock_NestingMode {
  INVALID = 0,
  SINGLE = 1,
  LIST = 2,
  SET = 3,
  MAP = 4,
  GROUP = 5,
  UNRECOGNIZED = -1,
}
// ...
export interface Schema_Object {
  attributes: Schema_Attribute[];
  nesting: Schema_Object_NestingMode;
  /**
   * MinItems and MaxItems were never used in the protocol, and have no
   * effect on validation.
   *
   * @deprecated
   */
  minItems: Long;
  /** @deprecated */
  maxItems: Long;
}

export enum Schema_Object_NestingMode {
  INVALID = 0,
  SINGLE = 1,
  LIST = 2,
  SET = 3,
  MAP = 4,
  UNRECOGNIZED = -1,
}
// ...
export const Schema: MessageFns<Schema> = {
  encode(message: Schema, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (!message.version.equals(Long.ZERO)) {
      writer.uint32(8).int64(message.version.toString());
    }
    if (message.block !== undefined) {
      Schema_Block.encode(message.block, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Schema {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSchema();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.version = Long.fromString(reader.int64().toString());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.block = Schema_Block.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Schema {
    return {
      version: isSet(object.version) ? Long.fromValue(object.version) : Long.ZERO,
      block: isSet(object.block) ? Schema_Block.fromJSON(object.block) : undefined,
    };
  },

  toJSON(message: Schema): unknown {
    const obj: any = {};
    if (!message.version.equals(Long.ZERO)) {
      obj.version = (message.version || Long.ZERO).toString();
    }
    if (message.block !== undefined) {
      obj.block = Schema_Block.toJSON(message.block);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Schema>, I>>(base?: I): Schema {
    return Schema.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Schema>, I>>(object: I): Schema {
    const message = createBaseSchema();
    message.version = (object.version !== undefined && object.version !== null)
      ? Long.fromValue(object.version)
      : Long.ZERO;
    message.block = (object.block !== undefined && object.block !== null)
      ? Schema_Block.fromPartial(object.block)
      : undefined;
    return message;
  },
};
// ...
export const Schema_NestedBlock: MessageFns<Schema_NestedBlock> = {
  encode(message: Schema_NestedBlock, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.typeName !== "") {
      writer.uint32(10).string(message.typeName);
    }
    if (message.block !== undefined) {
      Schema_Block.encode(message.block, writer.uint32(18).fork()).join();
    }
    if (message.nesting !== 0) {
      writer.uint32(24).int32(message.nesting);
    }
    if (!message.minItems.equals(Long.ZERO)) {
      writer.uint32(32).int64(message.minItems.toString());
    }
    if (!message.maxItems.equals(Long.ZERO)) {
      writer.uint32(40).int64(message.maxItems.toString());
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Schema_NestedBlock {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSchema_NestedBlock();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.typeName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.block = Schema_Block.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.nesting = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.minItems = Long.fromString(reader.int64().toString());
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.maxItems = Long.fromString(reader.int64().toString());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Schema_NestedBlock {
    return {
      typeName: isSet(object.typeName)
        ? globalThis.String(object.typeName)
        : isSet(object.type_name)
        ? globalThis.String(object.type_name)
        : "",
      block: isSet(object.block) ? Schema_Block.fromJSON(object.block) : undefined,
      nesting: isSet(object.nesting) ? schema_NestedBlock_NestingModeFromJSON(object.nesting) : 0,
      minItems: isSet(object.minItems)
        ? Long.fromValue(object.minItems)
        : isSet(object.min_items)
        ? Long.fromValue(object.min_items)
        : Long.ZERO,
      maxItems: isSet(object.maxItems)
        ? Long.fromValue(object.maxItems)
        : isSet(object.max_items)
        ? Long.fromValue(object.max_items)
        : Long.ZERO,
    };
  },

  toJSON(message: Schema_NestedBlock): unknown {
    const obj: any = {};
    if (message.typeName !== "") {
      obj.typeName = message.typeName;
    }
    if (message.block !== undefined) {
      obj.block = Schema_Block.toJSON(message.block);
    }
    if (message.nesting !== 0) {
      obj.nesting = schema_NestedBlock_NestingModeToJSON(message.nesting);
    }
    if (!message.minItems.equals(Long.ZERO)) {
      obj.minItems = (message.minItems || Long.ZERO).toString();
    }
    if (!message.maxItems.equals(Long.ZERO)) {
      obj.maxItems = (message.maxItems || Long.ZERO).toString();
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Schema_NestedBlock>, I>>(base?: I): Schema_NestedBlock {
    return Schema_NestedBlock.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Schema_NestedBlock>, I>>(object: I): Schema_NestedBlock {
    const message = createBaseSchema_NestedBlock();
    message.typeName = object.typeName ?? "";
    message.block = (object.block !== undefined && object.block !== null)
      ? Schema_Block.fromPartial(object.block)
      : undefined;
    message.nesting = object.nesting ?? 0;
    message.minItems = (object.minItems !== undefined && object.minItems !== null)
      ? Long.fromValue(object.minItems)
      : Long.ZERO;
    message.maxItems = (object.maxItems !== undefined && object.maxItems !== null)
      ? Long.fromValue(object.maxItems)
      : Long.ZERO;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends Long ? string | number | Long : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}